A  to-do list that can be modified and saved to a text file.
Inputs:
Text file to open for reading
User choice: (‘1’-Insert new to-do item, ‘2’-Modify existing to-do item,
‘3’- Remove existing to-do item, ‘4’-Generate list of to-do items,
‘5’-quit)
Output:
Updated list of to-do items, error messages when appropriate
Text file written with updated contents
Specification:
The VBScript program maintains a list of to-do items supplied by a
system administrator. It will be called by an HTML file in order to make
it portable and platform independent (although it only runs under
compatibility with Internet Explorer version 9 and earlier).
Input and output will be handled via dialog boxes (inputbox and
messagebox, respectively). Each option should be implemented as a
separate function within the VBScript file, and selected upon the
appropriate choice by the user.
The to-do items will be loaded from an input file and stored in a global
array which will be recognized and modified appropriately by each
function.
Initially, the program prompts for an input text file to read from. If
the file does not exist, it should be created. The program should then
loop until the user chooses to quit. If the user enters an illegal
status, the program will prompt again for the status input. Upon
quitting, the program prompts the user to save the list to an output
file with the same name as the original input text file.